requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
numpy==2.1.1
faiss-cpu==1.9.0
sentence-transformers==3.1.1
streamlit==1.38.0
kagglehub==0.3.13
pypdf==5.0.1
json-repair==0.33.0
tqdm==4.66.5
